+++ R/group_standards.R
@@ -118,7 +118,7 @@
                                                       max_nstd = max_nstd,
                                                       min_diff = min_diff)
         if (i > n*n)
-            stop("all combination were tested, no possibility to fit your input requirement")
+            stop("all combination were tested, no possibility to fit your input requirements")
     }  
     standard_groups
 }